more updates
making a function that inserts user transactions in a list (Done)
making a function refund request list (Done)
Refunding  approved request amount to user's wallet   (Not done)

+++ FinalTry/src/trial/etisalatMS.java
@@ -12,7 +12,7 @@ public class etisalatMS extends mobileServiceProvider {
          return var;
 	}
 	@Override
-	public void createForm(form f,service serviceName,double number,double money) {
+	public void createForm(form f,service serviceName,int number,double money) {
 		textField t2 = new textField();
 		dropDownField t1= new dropDownField();
 		f.Addfields(t2);
