+++ FinalTry/src/trial/weMS.java
@@ -15,7 +15,7 @@ public class weMS extends mobileServiceProvider{
 	}
 
 	@Override
-	public void createForm(form f,service serviceName,double number,double money) {
+	public void createForm(form f,service serviceName,int number,double money) {
 		textField t2 = new textField();
 		dropDownField t1= new dropDownField();
 		f.Addfields(t2);
